Split backfill types into one-type-per-file #597
Move BackfillResult, CurrencyPair, FxDailyStatus and FxDailyResult out of the
interface files they were declared in and into their own files, matching the
project convention of one non-nested type per file. No behaviour change.
