+++ code/FinanceManager.Application/Backfill/Currencies/CurrencyPair.cs
@@ -1,0 +1,4 @@
+namespace FinanceManager.Application.Backfill.Currencies;
+
+/// <summary>A normalized, upper-cased currency pair to backfill (<see cref="From"/> → <see cref="To"/>).</summary>
+public readonly record struct CurrencyPair(string From, string To);
